arch/delays: Set invalid default value for BOARD_LOOPSPERMSEC
This commit sets the default Kconfig value for CONFIG_BOARD_LOOPSPERMSEC
as -1, which is invalid. It statically asserts at compile time that the
value is greater than or equal to 0 so that the user is notified if
their board/configuration does not yet have a calibrated value for this
configuration option. The user is directed to set a non-negative value
and then later calibrate with `calib_udelay`.

QEMU ARMv8-A configurations were given a value obtained by running
`calib_udelay` on the NSH configuration.

Signed-off-by: Matteo Golin <matteo.golin@gmail.com>